requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121

opencv-python-headless>=4.8.0
basicsr-fixed==1.4.2
facexlib>=0.2.0
gfpgan>=1.3.8
huggingface_hub>=0.21.4
scikit-image>=0.22.0
tqdm>=4.66.0
pytest>=8.0.0
